django/requirements.txt
Django==2.0.2
django-cors-headers==2.1.0
django-extensions==1.9.9
djangorestframework==3.7.7
psycopg2==2.7.3.2
